+++ viewdata.js
@@ -42,6 +42,7 @@ function getViewData() {
                 "type": "TABLE"
             };
             propObj.display_value = "<a class='referencelink' href='" + url + "/" + tableName + ".do?sys_id=" + sysId + "' target='_blank'>" + tableName + " / " + sysId + "</a>";
+            propObj.link = url + "/" + tableName + ".do?sys_id=" + sysId;
             propObj.value = tableName + " / " + sysId;
             dataView.push(propObj);
 
@@ -87,6 +88,11 @@ function setViewData(nme) {
     Object.entries(nme).forEach( // add check of empty fields to be able to filter out
         ([key, obj]) => {
             nme[key].hasdata = (obj.value || obj.display_value) ? "hasdata" : "";
+            if (!nme[key].link ) {
+                let escpd = escape(nme[key].display_value);
+                if (nme[key].display_value != escpd)
+                    nme[key].display_value = '<pre style="white-space: pre-wrap; max-width:600px;"><code>' + escpd + '</code></pre>'
+            }
         }
     );
 
@@ -195,4 +201,12 @@ function getUrlVars(key) {
     return vars[key];
 }
 
+function escape(htmlStr) {
+    return htmlStr.replace(/&/g, "&amp;")
+          .replace(/</g, "&lt;")
+          .replace(/>/g, "&gt;")
+          .replace(/"/g, "&quot;")
+          .replace(/'/g, "&#39;");        
+ 
+ }