+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # CHANGELOG.md
 
-## 7.0.8.0 (2023-07-26 Firefox only)
+## 7.0.8.3 (2023-07-28)
+Fixes / changes:
+ - Escape display value in View Data (Issue: #421)
+ - Added function to scriptsync page to set banner message from sn-scriptsync
+
+## 7.0.8.1 (2023-07-26 Firefox only)
 Fixes / changes:
  - Workaround for issue #420 releated to Firefox containers compatability
  - Added wildcard in scriptsync condition